feature: Make statistics O(1)
Create a new object in the uploads cache to store an instance of the
case class "UploadStatistics" and create a thread worker that updates it
every second.

The endpoint then reads from the cached value always, making it O(1) in
time and "space" since it returns the result of the aggregation
performed previously in the background.

The code/approach is not very clean yet, lots of misplaced methods and
lack of proper abstractions, but I am following a baby-step approach
since I don't have much knowledge on Scala.

More about the approach will be available in the COMMENTS.md file (soon
to be added).